Customers Churn Segmentation use ANN/deployment/requirements.txt
streamlit
scikit-learn
pandas
numpy
tensorflow-cpu
